+++ src/main/java/com/giantlink/grh/services/impl/ProjectServiceImpl.java
@@ -1,12 +1,17 @@
 package com.giantlink.grh.services.impl;
 
 import com.giantlink.grh.entities.Project;
+import com.giantlink.grh.exceptions.NotFoundException;
+import com.giantlink.grh.mappers.ProjectMapper;
+import com.giantlink.grh.models.Requests.ProjetRequest;
+import com.giantlink.grh.models.Responses.ProjectResponse;
 import com.giantlink.grh.repositories.ProjectRepository;
 import com.giantlink.grh.services.ProjectService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProjectServiceImpl implements ProjectService {
@@ -18,22 +23,46 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
-    public List<Project> get() {
-        return projectRepository.findAll();
+    public List<ProjectResponse> get() throws NotFoundException {
+        List<Project> findProject = projectRepository.findAll();
+        if(findProject.isEmpty()){
+            throw new NotFoundException("No Project found");
+        }
+        return ProjectMapper.INSTANCE.projectListToResponse(findProject);
     }
 
     @Override
-    public Project get(Integer id) {
-        return projectRepository.findById(id).get();
+    public ProjectResponse get(Integer id) throws NotFoundException {
+        Optional<Project> findProjectId = projectRepository.findById(id);
+        if(!findProjectId.isPresent()){
+            throw new NotFoundException("Project with id : "+ id+ " not found");
+        }
+        return ProjectMapper.INSTANCE.projectToResponse(findProjectId.get());
     }
 
     @Override
-    public Project add (Project project) {
-        return projectRepository.save(project);
+    public ProjectResponse add (ProjetRequest projectRequest) {
+        Project project = ProjectMapper.INSTANCE.requestToProject(projectRequest);
+        return ProjectMapper.INSTANCE.projectToResponse(projectRepository.save(project));
     }
 
     @Override
-    public void delete(Integer id) {
+    public ProjectResponse update(Integer id, ProjetRequest projetRequest) throws NotFoundException{
+        Project project = ProjectMapper.INSTANCE.requestToProject(projetRequest);
+        if(!projectRepository.findById(id).isPresent()){
+            throw new NotFoundException("Project with id : "+ id+ " not found");
+        }
+        if(projectRepository.findById(id).isPresent()){
+            project.setId(id);
+        }
+        return ProjectMapper.INSTANCE.projectToResponse(project);
+    }
+
+    @Override
+    public void delete(Integer id) throws NotFoundException{
+        if(!projectRepository.findById(id).isPresent()){
+            throw new NotFoundException("Project with id : "+ id+ " not found");
+        }
         projectRepository.deleteById(id);
     }
 }
